% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_coherence.R
\name{estimate_coherence}
\alias{estimate_coherence}
\title{Coherence}
\usage{
estimate_coherence(y, x, x.no_inter)
}
\arguments{
\item{y}{Treatment Effect to be studied.}

\item{x}{Variables with interactions.}

\item{x.no_inter}{Variables without interactions.}
}
\value{
The value of coherence.
}
\description{
This function calculates the value of coherence for a model.
}
\section{Dependencies}{

 \itemize{
   \item{glmnet}
 }
}

\examples{
 estimate_coherence(
   y = df$teffect,
   x = df[, ml_variables],
   x.no_inter = df[[ml_variables[grep("^[^X]", ml_variables)]]]
   )
}
\keyword{Coherence}
