% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_main_effect.R
\name{estimate_main_effect}
\alias{estimate_main_effect}
\title{Estimate Main Effect}
\usage{
estimate_main_effect(y, treatment, X)
}
\arguments{
\item{y}{y-variable for the second lasso.}

\item{treatment}{Treatment to be studied.}

\item{X}{X-variables.}
}
\value{
The regression coefficients.
}
\description{
This function estimates the main effect of the treatment.
}
\examples{
 estimate_main_effect(
   y = cb_startups$equity_growth,
   treatment = cb_startups$bearly_stage_has_vc,
   X = cb_startups[, ml_vars]
   )
}
\keyword{Effect}
\keyword{Main}
