% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_propensity.R
\name{estimate_propensity}
\alias{estimate_propensity}
\title{Estimate Propensity Score}
\usage{
estimate_propensity(treatment, X)
}
\arguments{
\item{treatment}{Treatment to be studied.}

\item{X}{Complete sample of the x-variables to estimate the propensity score for all the observations in the dataframe.}
}
\value{
The propensity scores for each row in our dataset.
}
\description{
This function estimates the propensity score using a RandomForest model.
}
\section{Dependencies}{
 
 \describe {
   \item{randomForest}
   \item{dplyr}
 }
}

\examples{
 estimate_propensity(
   treatment = df$treatment,
   X = df[[ml_variables]]
   )
}
\keyword{Propensity}
\keyword{Score}
