% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ste.R
\name{estimate_ste}
\alias{estimate_ste}
\title{Estimate the Strategic Treatment Effect.}
\usage{
estimate_ste(
  y,
  treatment,
  propensity,
  df,
  span_treated = 0.75,
  span_untreated = 0.75
)
}
\arguments{
\item{y}{Outcome variable to be studied.}

\item{treatment}{Treatment variable.}

\item{propensity}{Propensity score generated earlier.}

\item{df}{The complete dataset.}
}
\value{
df The same as input dataframe with the added column of the strategic treatment effect.
}
\description{
This function estimates the strategic treatment effect.
}
\section{Dependencies}{

 \itemize{
   \item{dplyr}
 }
}

\examples{
estimate_ste(
   y = df$equity_growth,
   treatment = df$treatment,
   propensity = p_scores,
   df = df
)
}
\keyword{Effect}
\keyword{Strategic}
\keyword{Treatment}
