% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_determinants.R
\name{get_top_ste_determinants}
\alias{get_top_ste_determinants}
\title{Study Determinants of the STE.}
\usage{
get_top_ste_determinants(ste, X, teffect)
}
\arguments{
\item{ste}{Strategic Treatment Effect calculated earlier.}

\item{X}{x-variables for the LASSO.}

\item{teffect}{Treatment effect of the model.}
}
\value{
ste_features A dataframe containing the STE features.
}
\description{
This function studies the determinants of the strategic treatment effect.
}
\section{Dependencies}{

 \itemize{
   \item{glmnet}
   \item{dplyr}
 }
}

\examples{
get_top_ste_determinants(
   ste = df$ste,
   X = df[, ml_variables],
   teffect = df$teffect
)
}
\keyword{Determinants}
\keyword{STE}
