% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDBMS.R
\name{RDBMS}
\alias{RDBMS}
\title{a DB connector class}
\usage{
RDBMS(RDBMS.type = "mysql", user = "root", password = "",
  host = "127.0.0.1", dbname)
}
\arguments{
\item{RDBMS.type}{the RDBMS type. By default 'mysql', at the moment (the only supported)}

\item{user}{the user for the RDBMS connection}

\item{password}{the password for the RDBMS connection}

\item{host}{the host for the RDBMS connection}

\item{dbname}{the dataBase name for the RDBMS connection}
}
\description{
A class to connect with RDBMS
}

