% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_main_effect.R
\name{estimate_main_effect}
\alias{estimate_main_effect}
\title{Estimate Main Effect}
\usage{
estimate_main_effect(y_var, treatment_var, X, data_df)
}
\arguments{
\item{y_var}{String containing the name of the outcome variable to be studied.}

\item{treatment_var}{String containing the name of treatment variable to be studied.}

\item{X}{X-variables.}

\item{data_df}{The complete dataframe.}
}
\value{
The regression output from the fixed-effects OLS regression.
}
\description{
This function estimates the main effect of the treatment.
}
\section{Dependencies}{

 \itemize{
   \item{glmnet}
   \item{stringr}
   \item{fixest}
   \item{dplyr}
 }
}

\examples{
 estimate_main_effect(
   y_var = "outcome_var_name",
   treatment_var = "treatment_var_name",
   X = df[, ml_variables],
   data_df = df
   )
}
\keyword{Effect}
\keyword{Main}
